// src/core-node/infrastructure/enums/EHttpStatus.ts
export enum EHttpStatus {
    Success = 200,
    Created = 201,
    NotContent = 204,
    InvalidRequest = 400,
    Unauthorized = 401,
    Forbidden = 403,
    NotFound = 404,
    NotAcceptable = 406,
    RequestTimeout = 408,
    Conflict = 409,
    PreconditionRequired = 428,
    RunTimeError = 500,
    UknownError = 500,
    BadGateway = 502,
    ServiceUnavailable = 503,
    GatewayTimeout = 504,
}