// src/core-node/infrastructure/enums/EErrorMessage.ts
export enum EErrorMessages {
    DEFAULT_SCHEMA_ERROR = 'Ocurrió un error con las propiedades',
    CLIENT_ERROR = 'En estos momentos no lo podemos atender, favor intente más tarde',
    USER_AND_PASSWORD_ERROR = 'Usuario o contraseña incorrectos',
    SET_PROPERTY_ERROR = 'Se encontró un error con el parámetro :param',
    SEND_ERROR = 'Error al ejecutar el motor de reglas',
    USER_ERROR = 'Dispositivo no se encuentra registrado',
    NO_CONTENT = 'NO_CONTENT',
    NOT_FOUND = 'NOT_FOUND',
    NOT_CREATE = 'Registro no pudo ser ingresado, favor intente más tarde',
}